+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "",
+  "author": "Chris Flowers",
   "license": "ISC",
   "dependencies": {
     "request": "^2.75.0",
